Final commit

+++ src/main/java/com/hansen/sqlitetopostgres/repo/postgres/TableBRepository.java
@@ -2,7 +2,6 @@ package com.hansen.sqlitetopostgres.repo.postgres;
 
 import com.hansen.sqlitetopostgres.entity.postgres.TableB;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-public interface TableBRepository extends JpaRepository<TableB, Integer> {
+public interface TableBRepository extends JpaRepository<TableB, Long> {
 }
